robot_web_ui/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
pyserial==3.5

# LeRobot (使用本地安装)
-e /Users/ai/Project/lerobot
